requirements.txt
playwright==1.40.0
pandas==2.1.3
openpyxl==3.1.2
python-dateutil==2.8.2
tenacity==8.2.3
python-dotenv==1.0.0